+++ dflib/src/main/java/org/dflib/slice/AllRowColumnSet.java
@@ -10,6 +10,12 @@ import org.dflib.RowToValueMapper;
 import java.util.Map;
 import java.util.function.UnaryOperator;
 
+/**
+ * A {@link RowColumnSet} over the entire DataFrame with no filters, expansions or other row changes.
+ *
+ * @deprecated unused
+ */
+@Deprecated(since = "2.0.0", forRemoval = true)
 public class AllRowColumnSet implements RowColumnSet {
 
     private final ColumnSet columnSet;